pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["kuznetsovyar22 <kuznetsovyar@mail.ru>"]
readme = "README.md"
packages = [{include = "python_project_83"}]

[tool.poetry.dependencies]
python = "^3.8.2"
flask = "^2.2.3"
flake8 = "^6.0.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
datetime = "^5.1"
requests = "^2.28.2"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
